backend/requirements.txt
typer
wasabi
pathlib
black
fastapi
uvicorn
openai
weaviate-client==3.22.0
mypy
python-dotenv